feat: middleware 및 page.tsx => component화
