+++ src/components/LogoutBtn.tsx
@@ -1,0 +1,16 @@
+"use client";
+import { FiLogOut } from "react-icons/fi";
+import useLogOut from "@/hooks/useLogOut";
+
+export default function LogoutBtn() {
+  const { handleLogout } = useLogOut();
+
+  const handleLogOutClick = async () => {
+    await handleLogout();
+  };
+  return (
+    <button onClick={handleLogOutClick} className="mx-3">
+      <FiLogOut className="w-8 h-8 text-white" />
+    </button>
+  );
+}
